chore: Extract `useAhbapData` hook
